修复当一个应用没有cmd_set_def时增加按钮对Admin和PE不显示的bug。
git-svn-id: https://apposs-center.googlecode.com/svn/trunk@19 e3cf8a15-5072-d663-4c44-a26c1ac8da71

+++ app/controllers/cmd_set_defs_controller.rb
@@ -3,14 +3,19 @@ class CmdSetDefsController < BaseController
     app_id = params[:app_id]
     role_id = Stakeholder.where(:app_id => app_id, :user_id => current_user.id).first.role_id
     role = Role.find(role_id).name
+    cmd_set_defs = current_app.cmd_set_defs
     if role == Role::Admin || role == Role::PE
-      respond_with current_app.cmd_set_defs.collect { |obj|
-        obj.serializable_hash.update("actions" => [
-            {:name=>"准备执行", :flex => 1.7, :url=> app_cmd_sets_path(app_id), :type => 'simple', :method => 'POST'},
-            {:name=>"修改", :flex => 1, :url=> edit_app_cmd_set_def_path(app_id, obj.id), :type => 'multi', :method => 'GET'},
-            {:name=>"删除", :flex => 1, :url=> app_cmd_set_def_path(app_id, obj.id), :type => 'delete', :method => 'DELETE'}
-        ], "flex" => 8, "add" => true)
-      }
+      if cmd_set_defs.length == 0
+        respond_with [{"add" => true}]
+      else
+        respond_with cmd_set_defs.collect { |obj|
+          obj.serializable_hash.update("actions" => [
+              {:name=>"准备执行", :flex => 1.7, :url=> app_cmd_sets_path(app_id), :type => 'simple', :method => 'POST'},
+              {:name=>"修改", :flex => 1, :url=> edit_app_cmd_set_def_path(app_id, obj.id), :type => 'multi', :method => 'GET'},
+              {:name=>"删除", :flex => 1, :url=> app_cmd_set_def_path(app_id, obj.id), :type => 'delete', :method => 'DELETE'}
+          ], "flex" => 8, "add" => true)
+        }
+      end
     else
       respond_with current_app.cmd_set_defs.collect { |obj|
         obj.serializable_hash.update("actions" => [
